refactor(memory): decouple memory presenter
