+++ scripts/architecture-guard.mjs
@@ -35,6 +35,23 @@ const RENDERER_TYPED_BOUNDARY_WINDOW_API_ALLOWLIST = [
   path.join(ROOT, 'src/renderer/api/runtime.ts')
 ]
 const MAIN_SOURCE_ROOT = path.join(ROOT, 'src/main')
+const MEMORY_PRESENTER_ROOT = path.join(ROOT, 'src/main/presenter/memoryPresenter')
+const MEMORY_PRESENTER_FACADE_PATH = path.join(MEMORY_PRESENTER_ROOT, 'index.ts')
+const MEMORY_PRESENTER_CORE_ROOT = path.join(MEMORY_PRESENTER_ROOT, 'core')
+const MEMORY_PRESENTER_INFRA_ROOT = path.join(MEMORY_PRESENTER_ROOT, 'infra')
+const MEMORY_PRESENTER_SERVICES_ROOT = path.join(MEMORY_PRESENTER_ROOT, 'services')
+const MEMORY_PRESENTER_CORE_ALLOWED_ROOT_MODULES = new Set([
+  path.join(MEMORY_PRESENTER_ROOT, 'types.ts')
+])
+const MEMORY_PRESENTER_RUNTIME_ALLOWED_ROOT_MODULES = new Set([
+  path.join(MEMORY_PRESENTER_ROOT, 'context.ts'),
+  path.join(MEMORY_PRESENTER_ROOT, 'ports.ts'),
+  path.join(MEMORY_PRESENTER_ROOT, 'runtimeConstants.ts'),
+  path.join(MEMORY_PRESENTER_ROOT, 'types.ts')
+])
+const MEMORY_PRESENTER_SERVICE_LEAF_MODULES = new Set([
+  path.join(MEMORY_PRESENTER_SERVICES_ROOT, 'rowMutations.ts')
+])
 const PHASE_ORDER = new Map([
   ['P0', 0],
   ['P1', 1],
@@ -239,6 +256,100 @@ async function collectHotPathDirectEdges() {
   return edges.sort()
 }
 
+function memoryPresenterLayer(filePath) {
+  if (!isUnder(filePath, MEMORY_PRESENTER_ROOT)) return null
+  if (isUnder(filePath, MEMORY_PRESENTER_CORE_ROOT)) return 'core'
+  if (isUnder(filePath, MEMORY_PRESENTER_INFRA_ROOT)) return 'infra'
+  if (isUnder(filePath, MEMORY_PRESENTER_SERVICES_ROOT)) return 'services'
+  return 'root'
+}
+
+function isAllowedMemoryPresenterCoreRootModule(filePath) {
+  return MEMORY_PRESENTER_CORE_ALLOWED_ROOT_MODULES.has(path.resolve(filePath))
+}
+
+function isAllowedMemoryPresenterRuntimeRootModule(filePath) {
+  return MEMORY_PRESENTER_RUNTIME_ALLOWED_ROOT_MODULES.has(path.resolve(filePath))
+}
+
+function isMemoryPresenterFacade(filePath) {
+  return path.resolve(filePath) === path.resolve(MEMORY_PRESENTER_FACADE_PATH)
+}
+
+async function checkMemoryPresenterLayerImports(filePath, specifiers, violations) {
+  const importerLayer = memoryPresenterLayer(filePath)
+  if (!importerLayer) return
+
+  for (const specifier of specifiers) {
+    const resolved = await resolveImport(specifier, filePath)
+    if (!resolved || !isUnder(resolved, MEMORY_PRESENTER_ROOT)) continue
+
+    const importedLayer = memoryPresenterLayer(resolved)
+    if (!importedLayer) continue
+
+    if (importerLayer === 'root') {
+      if (isMemoryPresenterFacade(filePath)) continue
+
+      const allowed =
+        importedLayer === 'core' || (importedLayer === 'root' && !isMemoryPresenterFacade(resolved))
+      if (!allowed) {
+        violations.push(
+          `[memory-presenter-layer] ${relativePath(filePath)} -> ${relativePath(resolved)}; only memoryPresenter/index.ts may import services, infra, or the facade entrypoint`
+        )
+      }
+      continue
+    }
+
+    if (importerLayer === 'core') {
+      const allowed = importedLayer === 'core' || isAllowedMemoryPresenterCoreRootModule(resolved)
+      if (!allowed) {
+        violations.push(
+          `[memory-presenter-layer] ${relativePath(filePath)} -> ${relativePath(resolved)}; core may only import core files and root contracts`
+        )
+      }
+      continue
+    }
+
+    if (importerLayer === 'infra') {
+      const allowed =
+        importedLayer === 'infra' ||
+        importedLayer === 'core' ||
+        isAllowedMemoryPresenterRuntimeRootModule(resolved)
+      if (!allowed) {
+        violations.push(
+          `[memory-presenter-layer] ${relativePath(filePath)} -> ${relativePath(resolved)}; infra must not import services or facade entrypoints`
+        )
+      }
+      continue
+    }
+
+    if (importerLayer === 'services') {
+      const sameFile = path.resolve(filePath) === path.resolve(resolved)
+      const allowedServiceLeaf = MEMORY_PRESENTER_SERVICE_LEAF_MODULES.has(path.resolve(resolved))
+      const allowedRootModule =
+        importedLayer !== 'root' || isAllowedMemoryPresenterRuntimeRootModule(resolved)
+
+      if (importedLayer === 'services' && !sameFile && !allowedServiceLeaf) {
+        violations.push(
+          `[memory-presenter-layer] ${relativePath(filePath)} -> ${relativePath(resolved)}; service-to-service imports must use facade ports, except rowMutations`
+        )
+      }
+
+      if (importedLayer === 'infra') {
+        violations.push(
+          `[memory-presenter-layer] ${relativePath(filePath)} -> ${relativePath(resolved)}; services must depend on root port contracts, not infra concrete modules`
+        )
+      }
+
+      if (!allowedRootModule) {
+        violations.push(
+          `[memory-presenter-layer] ${relativePath(filePath)} -> ${relativePath(resolved)}; services may only import root runtime contracts`
+        )
+      }
+    }
+  }
+}
+
 async function loadBridgeRegister() {
   const raw = await fs.readFile(BRIDGE_REGISTER_PATH, 'utf8')
   const parsed = JSON.parse(raw)
@@ -359,6 +470,8 @@ async function main() {
   for (const filePath of [...fileSet].sort()) {
     const source = await fs.readFile(filePath, 'utf8')
     const specifiers = extractModuleSpecifiers(source)
+
+    await checkMemoryPresenterLayerImports(filePath, specifiers, violations)
 
     if (RENDERER_BUSINESS_ROOTS.some((root) => isUnder(filePath, root))) {
       const file = relativePath(filePath)
