+++ src/main/presenter/index.ts
@@ -69,7 +69,7 @@ import { ScheduledTasksService } from './scheduledTasks'
 import { AgentSessionPresenter } from './agentSessionPresenter'
 import { AgentRuntimePresenter } from './agentRuntimePresenter'
 import { MemoryPresenter, isSafeAgentId } from './memoryPresenter'
-import { MemoryVectorStore } from './memoryPresenter/memoryVectorStore'
+import { MemoryVectorStore } from './memoryPresenter/infra/memoryVectorStore'
 import { ProjectPresenter } from './projectPresenter'
 import { RemoteControlPresenter } from './remoteControlPresenter'
 import type { RemoteControlPresenterLike } from './remoteControlPresenter/interface'
