--- src/main/presenter/memoryPresenter/lifecycleConstants.ts
@@ -1,5 +1,0 @@
-export const ARCHIVE_DECAY_THRESHOLD = 0.05
-export const ARCHIVE_AGE_MS = 90 * 24 * 60 * 60 * 1000
-
-// Display-only band boundary: >= fresh, [archive threshold, fresh) aging, below archive stale.
-export const FRESH_DECAY_THRESHOLD = 0.5
