+++ src/main/presenter/memoryPresenter/runtimeConstants.ts
@@ -1,0 +1,40 @@
+import { MEMORY_HEALTH_DEFAULT_AUDIT_SCAN_LIMIT } from '@shared/contracts/routes/memory.routes'
+
+export const MIN_MEMORIES_FOR_REFLECTION = 3
+export const REFLECTION_IMPORTANCE_THRESHOLD = 5.0
+export const REFLECTION_IMPORTANCE = 0.8
+export const REFLECTION_MEMORY_LIMIT = 20
+
+export const MIN_MEMORIES_FOR_PERSONA = 3
+export const PERSONA_EVOLUTION_IMPORTANCE_THRESHOLD = 5.0
+export const PERSONA_MEMORY_LIMIT = 20
+
+export const WORKING_BLOB_TOKEN_LIMIT = 400
+export const WORKING_PROVENANCE_SEED = 'session-working-blob'
+
+export const REINDEX_BATCH_SIZE = 50
+export const REINDEX_MAX_BATCHES = 200
+
+export const DECISION_NEIGHBOR_TOP_S = 10
+
+export const CONSOLIDATION_IDLE_MS = 5 * 60 * 1000
+export const CONSOLIDATION_COOLDOWN_MS = 6 * 60 * 60 * 1000
+export const CONSOLIDATION_MAX_LLM_CALLS = 8
+export const CONSOLIDATION_MAX_INPUT_TOKENS = 24000
+export const CONSOLIDATION_MERGE_SIMILARITY = 0.85
+
+export const MAINTENANCE_START_DELAY_MS = 60 * 1000
+export const STARTUP_PREWARM_DELAY_MS = 3 * 1000
+export const STARTUP_ARM_STAGGER_MS = 5 * 1000
+export const STARTUP_PREWARM_STAGGER_MS = 1500
+
+export const EMBEDDING_PREWARM_TEXT = 'memory warmup'
+export const WARM_DIMENSION_FAILURE_COOLDOWN_MS = 30 * 1000
+
+export const MEMORY_HEALTH_TOP_ACCESSED_LIMIT = 5
+export const MEMORY_HEALTH_AUDIT_SCAN_LIMIT = MEMORY_HEALTH_DEFAULT_AUDIT_SCAN_LIMIT
+export const MEMORY_HEALTH_RECENT_FAILURES_LIMIT = 5
+export const MEMORY_CREATED_IDS_EVENT_LIMIT = 50
+
+export const RECALL_QUERY_EMBEDDING_TIMEOUT_MS = 800
+export const RECALL_QUERY_EMBEDDING_STALE_MS = 30 * 1000
