+++ test/main/scripts/architectureGuard.test.ts
@@ -8,9 +8,36 @@ const FIXTURE_PATH = path.join(
   ROOT,
   'src/renderer/settings/__architecture_guard_legacy_fixture__.ts'
 )
+const MEMORY_CORE_FIXTURE_PATH = path.join(
+  ROOT,
+  'src/main/presenter/memoryPresenter/core/__architecture_guard_core_fixture__.ts'
+)
+const MEMORY_INFRA_FIXTURE_PATH = path.join(
+  ROOT,
+  'src/main/presenter/memoryPresenter/infra/__architecture_guard_infra_fixture__.ts'
+)
+const MEMORY_SERVICE_FIXTURE_PATH = path.join(
+  ROOT,
+  'src/main/presenter/memoryPresenter/services/__architecture_guard_service_fixture__.ts'
+)
+const MEMORY_ROOT_FIXTURE_PATH = path.join(
+  ROOT,
+  'src/main/presenter/memoryPresenter/__architecture_guard_root_fixture__.ts'
+)
+const FIXTURE_PATHS = [
+  FIXTURE_PATH,
+  MEMORY_CORE_FIXTURE_PATH,
+  MEMORY_INFRA_FIXTURE_PATH,
+  MEMORY_SERVICE_FIXTURE_PATH,
+  MEMORY_ROOT_FIXTURE_PATH
+]
 
 async function writeSettingsFixture(source: string) {
   await writeFile(FIXTURE_PATH, source, 'utf8')
+}
+
+async function writeFixture(filePath: string, source: string) {
+  await writeFile(filePath, source, 'utf8')
 }
 
 function runArchitectureGuard() {
@@ -22,7 +49,7 @@ function runArchitectureGuard() {
 
 describe.sequential('architecture guard', () => {
   afterEach(async () => {
-    await rm(FIXTURE_PATH, { force: true })
+    await Promise.all(FIXTURE_PATHS.map((filePath) => rm(filePath, { force: true })))
   })
 
   it('fails when settings imports or calls the retired legacy presenter bridge', async () => {
@@ -51,5 +78,85 @@ describe.sequential('architecture guard', () => {
     expect(result.status).not.toBe(0)
     expect(result.stderr).toContain('[renderer-business-direct-window-electron]')
     expect(result.stderr).toContain('[renderer-business-direct-ipc-listener]')
+  })
+
+  it('fails when memory core imports runtime context', async () => {
+    await writeFixture(
+      MEMORY_CORE_FIXTURE_PATH,
+      `
+        import type { MemoryRuntimeContext } from '../context'
+        export type Fixture = MemoryRuntimeContext
+      `
+    )
+
+    const result = runArchitectureGuard()
+
+    expect(result.status).not.toBe(0)
+    expect(result.stderr).toContain('[memory-presenter-layer]')
+    expect(result.stderr).toContain('core may only import core files and root contracts')
+  })
+
+  it('fails when memory infra imports services', async () => {
+    await writeFixture(
+      MEMORY_INFRA_FIXTURE_PATH,
+      `
+        import type { WorkingMemoryService } from '../services/workingMemoryService'
+        export type Fixture = WorkingMemoryService
+      `
+    )
+
+    const result = runArchitectureGuard()
+
+    expect(result.status).not.toBe(0)
+    expect(result.stderr).toContain('[memory-presenter-layer]')
+    expect(result.stderr).toContain('infra must not import services or facade entrypoints')
+  })
+
+  it('fails when memory services import another concrete service or infra concrete module', async () => {
+    await writeFixture(
+      MEMORY_SERVICE_FIXTURE_PATH,
+      `
+        import type { WorkingMemoryService } from './workingMemoryService'
+        import type { VectorStoreManager } from '../infra/vectorStoreManager'
+        export type Fixture = WorkingMemoryService | VectorStoreManager
+      `
+    )
+
+    const result = runArchitectureGuard()
+
+    expect(result.status).not.toBe(0)
+    expect(result.stderr).toContain('service-to-service imports must use facade ports')
+    expect(result.stderr).toContain('services must depend on root port contracts')
+  })
+
+  it('allows memory services to import the shared row mutation leaf', async () => {
+    await writeFixture(
+      MEMORY_SERVICE_FIXTURE_PATH,
+      `
+        import type { MemoryRowMutations } from './rowMutations'
+        export type Fixture = MemoryRowMutations
+      `
+    )
+
+    const result = runArchitectureGuard()
+
+    expect(result.status).toBe(0)
+  })
+
+  it('fails when non-facade memory root files import the facade or service layer', async () => {
+    await writeFixture(
+      MEMORY_ROOT_FIXTURE_PATH,
+      `
+        import type { MemoryPresenter } from './index'
+        import type { WorkingMemoryService } from './services/workingMemoryService'
+        export type Fixture = MemoryPresenter | WorkingMemoryService
+      `
+    )
+
+    const result = runArchitectureGuard()
+
+    expect(result.status).not.toBe(0)
+    expect(result.stderr).toContain('[memory-presenter-layer]')
+    expect(result.stderr).toContain('only memoryPresenter/index.ts may import services')
   })
 })
